refactor: create shared components
