+++ src/components/tests/test05/SpeechBubble.tsx
@@ -1,0 +1,14 @@
+type SpeechBubbleProps = {
+  showAnswer: () => void;
+};
+
+export default function SpeechBubble({ showAnswer }: SpeechBubbleProps) {
+  return (
+    <div className="speech-bubble-container" onClick={showAnswer}>
+      <div className="blur"></div>
+      <div className="speech-bubble-text">
+        <p>정답</p>
+      </div>
+    </div>
+  );
+}
