+++ src/pages/Home.tsx
@@ -1,5 +1,4 @@
-import { useState } from "react";
-import { Link } from "react-router-dom";
+import SelectTypeButton from "@components/common/SelectTypeButton";
 
 export default function Home() {
   const practiceType = [
@@ -20,7 +19,12 @@ export default function Home() {
             <p className="font-light">실행할 듣기 연습 종류를 선택하세요.</p>
             <ul className="select-type">
               {practiceType.map((type, i) => (
-                <SelectTypeButton key={i} to={type.path} children={type.name} />
+                <SelectTypeButton
+                  key={i}
+                  to={type.path}
+                  children={type.name}
+                  className=""
+                />
               ))}
             </ul>
           </div>
@@ -33,38 +37,5 @@ export default function Home() {
         </div>
       </div>
     </div>
-  );
-}
-
-function SelectTypeButton({ to, children }: any) {
-  const [isHovered, setIsHovered] = useState(false);
-  const [isClicked, setIsClicked] = useState(false);
-
-  const handleMouseEnter = () => setIsHovered(true);
-  const handleMouseLeave = () => setIsHovered(false);
-  const handleMouseDown = () => setIsClicked(true);
-  const handleMouseUp = () => setIsClicked(false);
-  const buttonStyle = {
-    backgroundColor: isClicked
-      ? "#40A0FF"
-      : isHovered
-      ? "rgba(99, 180, 255, 0.1)"
-      : "#fff",
-    color: isClicked ? "#fff" : "#4894fe",
-    border: isHovered ? "3px solid transparent" : "3px solid #4894fe",
-  };
-
-  return (
-    <Link
-      className="select-type__button"
-      to={to}
-      style={buttonStyle}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={handleMouseLeave}
-      onMouseDown={handleMouseDown}
-      onMouseUp={handleMouseUp}
-    >
-      {children}
-    </Link>
   );
 }
