+++ src/pages/tests/test01/Test01Menu.tsx
@@ -1,6 +1,7 @@
-import { Link } from "react-router-dom";
-import Snb from "@components/common/Snb";
 import categories from "@datas/testCategories.json";
+
+import Snb from "@components/common/Snb";
+import DifficultyCard from "@components/tests/test01/DifficultyCard";
 
 export default function Test01Menu() {
   const testCategories = categories["01"];
@@ -26,30 +27,6 @@ export default function Test01Menu() {
             ))}
           </div>
         </div>
-      </div>
-    </div>
-  );
-}
-
-function DifficultyCard({ showDifficulty, difficulty, noiseLevel, snr, to }: any) {
-  return (
-    <div className="test-card-wrapper">
-      <div className="test-card-left">
-        {showDifficulty && <p className="font-light">난이도</p>}
-        <p className="font-bold">{difficulty}</p>
-      </div>
-      <div className="test-card-right">
-        <div className="test-card-right__row">
-          <img
-            src={`${process.env.PUBLIC_URL}/images/icons/icon_bookmark.png`}
-            alt="Bookmark Icon"
-          />
-          <p className="font-bold margin">{noiseLevel}</p>
-          <p className="text-gray">{snr}</p>
-        </div>
-        <Link to={to} className="test-card-button">
-          이동하기
-        </Link>
       </div>
     </div>
   );
