+++ src/pages/tests/test02/Test02Level.tsx
@@ -1,8 +1,12 @@
 import { useEffect, useState } from "react";
 import { Link, useLocation, useParams } from "react-router-dom";
+
 import useAxios, { API_URL, IRequestType } from "@hooks/useAxios";
-import Snb from "@components/common/Snb";
+
 import categories from "@datas/testCategories.json";
+
+import Snb from "@components/common/Snb";
+import StatusCard from "@components/tests/StatusCard";
 
 const CARDS_PER_PAGE = 10;
 
@@ -74,34 +78,5 @@ export default function Test02Level() {
         </div>
       </div>
     </div>
-  );
-}
-
-interface StatusCardProps {
-  number: number;
-}
-function StatusCard({ number }: StatusCardProps) {
-  const [progress, setProgress] = useState(50);
-  const location = useLocation();
-
-  const testScreenPath = `${location.pathname}/${number}`;
-
-  return (
-    <Link to={testScreenPath} className="status-card">
-      <p className="status-card__number">{number}</p>
-      <div className="status-card__progress-bar">
-        <div
-          className="status-card__progress-fill"
-          style={{ width: `${progress}%` }}
-        ></div>
-      </div>
-      <div className="status-card__percentage">{progress}%</div>
-      <div className="status-card__complete">
-        <img
-          src={`${process.env.PUBLIC_URL}/images/icons/icon_progress_check_white.png`}
-          alt="White Check Icon"
-        />
-      </div>
-    </Link>
   );
 }
