+++ src/pages/tests/test02/Test02Screen.tsx
@@ -1,9 +1,14 @@
 import { useEffect, useState } from "react";
-import { Link, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
+
+import data from "@datas/test02Data.json";
+
 import Snb from "@components/common/Snb";
 import ToggleSwitch from "@components/common/ToggleSwitch";
 import AnswerButton from "@components/common/AnswerButton";
-import data from "@datas/test02Data.json";
+import SelectTypeButton from "@components/common/SelectTypeButton";
+import InteractiveTestButton from "@components/tests/InteractiveTestButton";
+import Pagination from "@components/tests/Pagination";
 
 export default function Test02Screen() {
   const [currentContext, setCurrentContext] = useState("");
@@ -140,21 +145,21 @@ export default function Test02Screen() {
               {!isFinished ? (
                 <>
                   <div className="test-contents__buttons row">
-                    <CustomButton
+                    <InteractiveTestButton
                       text="이야기 듣기"
                       defaultIcon={`${process.env.PUBLIC_URL}/images/icons/icon_speaker.png`}
                       blueIcon={`${process.env.PUBLIC_URL}/images/icons/icon_speaker_blue.png`}
                       whiteIcon={`${process.env.PUBLIC_URL}/images/icons/icon_speaker_white.png`}
                       onClick={handlePlayClick}
                     />
-                    <CustomButton
+                    <InteractiveTestButton
                       text="이야기 보기"
                       defaultIcon={`${process.env.PUBLIC_URL}/images/icons/icon_more.png`}
                       blueIcon={`${process.env.PUBLIC_URL}/images/icons/icon_more_blue.png`}
                       whiteIcon={`${process.env.PUBLIC_URL}/images/icons/icon_more_white.png`}
                       onClick={handleContextButtonClick}
                     />
-                    <CustomButton
+                    <InteractiveTestButton
                       text="정답 보기"
                       defaultIcon={`${process.env.PUBLIC_URL}/images/icons/icon_more.png`}
                       blueIcon={`${process.env.PUBLIC_URL}/images/icons/icon_more_blue.png`}
@@ -166,7 +171,11 @@ export default function Test02Screen() {
                     <div className="context__arrows">
                       <button onClick={handleLeftArrowClick}>
                         <img
-                          src={`${process.env.PUBLIC_URL}/images/icons/icon_arrow_left.png`}
+                          src={
+                            currentQuestionIndex === 1
+                              ? `${process.env.PUBLIC_URL}/images/icons/icon_arrow_left_disabled.png`
+                              : `${process.env.PUBLIC_URL}/images/icons/icon_arrow_left.png`
+                          }
                           alt="Left Arrow Icon"
                           className="left-arrow"
                           style={{
@@ -177,7 +186,11 @@ export default function Test02Screen() {
                       </button>
                       <button onClick={handleRightArrowClick}>
                         <img
-                          src={`${process.env.PUBLIC_URL}/images/icons/icon_arrow_right.png`}
+                          src={
+                            currentQuestionIndex === totalQuestions
+                              ? `${process.env.PUBLIC_URL}/images/icons/icon_arrow_right_disabled.png`
+                              : `${process.env.PUBLIC_URL}/images/icons/icon_arrow_right.png`
+                          }
                           alt="Right Arrow Icon"
                           className="right-arrow"
                           style={{
@@ -221,7 +234,11 @@ export default function Test02Screen() {
                   </div>
                 </>
               ) : (
-                <SelectTypeButton children="한번 더 연습하기" to={"/training/part2"} />
+                <SelectTypeButton
+                  children="한번 더 연습하기"
+                  to={"/training/part2"}
+                  className=""
+                />
               )}
             </div>
             {!isFinished && (
@@ -235,118 +252,6 @@ export default function Test02Screen() {
           </div>
         </div>
       </div>
-    </div>
-  );
-}
-
-interface ButtonProps {
-  text: string;
-  defaultIcon: string;
-  blueIcon: string;
-  whiteIcon: string;
-  onClick: () => void;
-}
-function CustomButton({ text, defaultIcon, blueIcon, whiteIcon, onClick }: ButtonProps) {
-  const [isHovered, setIsHovered] = useState(false);
-  const [isClicked, setIsClicked] = useState(false);
-
-  const handleButtonClick = () => {
-    setIsClicked(!isClicked);
-    onClick();
-  };
-
-  const backgroundColor = isClicked
-    ? "#40A0FF"
-    : isHovered
-    ? "rgba(99, 180, 255, 0.10)"
-    : "#fff";
-  const textColor = isClicked ? "#fff" : "#4894fe";
-  const iconSrc = isClicked ? whiteIcon : isHovered ? blueIcon : defaultIcon;
-  const imageStyle = {
-    transform: text !== "이야기 듣기" && isClicked ? "rotateX(180deg)" : "rotateX(0deg)",
-    transition: "transform 0.4s",
-  };
-
-  return (
-    <button
-      className="btn"
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-      onClick={handleButtonClick}
-      style={{
-        backgroundColor: backgroundColor,
-        color: textColor,
-      }}
-    >
-      <p>{text}</p>
-      <img src={iconSrc} alt={text} style={imageStyle} />
-    </button>
-  );
-}
-
-function SelectTypeButton({ to, children }: any) {
-  const [isHovered, setIsHovered] = useState(false);
-  const [isClicked, setIsClicked] = useState(false);
-
-  const handleMouseEnter = () => setIsHovered(true);
-  const handleMouseLeave = () => setIsHovered(false);
-  const handleMouseDown = () => setIsClicked(true);
-  const handleMouseUp = () => setIsClicked(false);
-  const buttonStyle = {
-    backgroundColor: isClicked
-      ? "#40A0FF"
-      : isHovered
-      ? "rgba(99, 180, 255, 0.1)"
-      : "#fff",
-    color: isClicked ? "#fff" : "#4894fe",
-    border: isHovered ? "3px solid transparent" : "3px solid #4894fe",
-  };
-
-  return (
-    <Link
-      className="select-type__button"
-      to={to}
-      style={buttonStyle}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={handleMouseLeave}
-      onMouseDown={handleMouseDown}
-      onMouseUp={handleMouseUp}
-    >
-      {children}
-    </Link>
-  );
-}
-
-interface PaginationProps {
-  currentPage: number;
-  totalPages: number;
-  onPageChange: (pageNumber: number) => void;
-  handleFinished: () => void;
-}
-function Pagination({
-  currentPage,
-  totalPages,
-  onPageChange,
-  handleFinished,
-}: PaginationProps) {
-  return (
-    <div className="pagination-wrapper">
-      <ul>
-        {[...Array(totalPages)].map((_, index) => {
-          const number = index + 1;
-          return (
-            <li
-              key={number}
-              className={`page-item ${number === currentPage ? "active" : ""}`}
-            >
-              <button onClick={() => onPageChange(number)}>{number}</button>
-            </li>
-          );
-        })}
-        <li className="finish-button">
-          <button onClick={handleFinished}>연습 마치기</button>
-        </li>
-      </ul>
     </div>
   );
 }
