+++ src/pages/tests/test03/Test03Level.tsx
@@ -1,8 +1,12 @@
 import { useEffect, useState } from "react";
 import { Link, useLocation, useParams } from "react-router-dom";
+
 import useAxios, { API_URL, IRequestType } from "@hooks/useAxios";
-import Snb from "@components/common/Snb";
+
 import categories from "@datas/testCategories.json";
+
+import Snb from "@components/common/Snb";
+import StatusCard from "@components/tests/StatusCard";
 
 const CARDS_PER_PAGE = 7;
 
@@ -74,26 +78,5 @@ export default function Test03Level() {
         </div>
       </div>
     </div>
-  );
-}
-
-interface StatusCardProps {
-  number: number;
-}
-function StatusCard({ number }: StatusCardProps) {
-  const location = useLocation();
-
-  const testScreenPath = `${location.pathname}/${number}`;
-
-  return (
-    <Link to={testScreenPath} className="status-card margin">
-      <p className="status-card__title">긴 이야기 {number}</p>
-      <div className="status-card__complete">
-        <img
-          src={`${process.env.PUBLIC_URL}/images/icons/icon_progress_check_white.png`}
-          alt="White Check Icon"
-        />
-      </div>
-    </Link>
   );
 }
